refactor(dev): remove "Plugin" suffix from shared css plugins
