+++ packages/remix-dev/compiler/serverjs/compiler.ts
@@ -8,10 +8,10 @@ import type { RemixConfig } from "../../config";
 import type { AssetsManifest } from "../assets";
 import { loaders } from "../loaders";
 import type { CompileOptions } from "../options";
-import { cssModulesPlugin } from "../plugins/cssModulesPlugin";
-import { cssSideEffectImportsPlugin } from "../plugins/cssSideEffectImportsPlugin";
+import { cssModulesPlugin } from "../plugins/cssModuleImports";
+import { cssSideEffectImportsPlugin } from "../plugins/cssSideEffectImports";
 import { vanillaExtractPlugin } from "../plugins/vanillaExtract";
-import { cssFilePlugin } from "../plugins/cssFilePlugin";
+import { cssFilePlugin } from "../plugins/cssImports";
 import { deprecatedRemixPackagePlugin } from "../plugins/deprecatedRemixPackagePlugin";
 import { emptyModulesPlugin } from "../plugins/emptyModulesPlugin";
 import { mdxPlugin } from "../plugins/mdx";
